fix: generate config maps for grafana
The Grafana container uses a config map as a volume for dashboards. Another container in the same pod watches the config-map and reloads the configuration for Grafana so that it refreshes the dashboards once a dashboard is added, removed or altered.

Had to do a kubectl replace instead of kubectl apply to replace the config-map since it grew too big for kubernetes to store a version of it in case we want to roll back. Kubernetes has a maximum size for versions of config maps. To circumvent this I had to use kubectl replace instead of apply. We have the dashboards in version controll so no point for kubernetes to version this config map.

Also, regenerated the config map so that it uses the dashboards that have been added and modified.